requirements.txt
torch
torchvision
numpy
pandas
tqdm
Pillow 
matplotlib
scipy
torchio
einops
scikit-learn
opencv-python